updating schema

+++ calculationservice/src/main/java/com/inmarket/interviews/calculationservice/graphql/Calculation.java
@@ -1,6 +1,4 @@
 package com.inmarket.interviews.calculationservice.graphql;
 
-import java.math.BigInteger;
-
-public record Calculation(BigInteger value) {
+public record Calculation(long value) {
 }
